Norminette correction
