+++ inc/execution.h
@@ -6,7 +6,7 @@
 /*   By: pabeckha <pabeckha@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 13:30:08 by pabeckha          #+#    #+#             */
-/*   Updated: 2024/04/18 16:18:52 by pabeckha         ###   ########.fr       */
+/*   Updated: 2024/04/18 19:45:34 by pabeckha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,5 +58,7 @@ void		store_last_path(t_info *structure);
 int			open_files(t_cmd *table);
 int			open_in_files(t_cmd *table);
 int			open_out_files(t_cmd *table);
+void		commands_execution(t_info *structure, int i);
+void	execute_child_process(t_info *structure, int i);
 
 #endif
